Revert "Security1 (#216)"
This reverts commit 9fad97a7b5298b2ececc4d2c710bba9363c665a5.

+++ lib/src/core/comms/security_header.rs
@@ -88,7 +88,8 @@ impl BinaryEncoder<AsymmetricSecurityHeader> for AsymmetricSecurityHeader {
 
         // validate sender_certificate_length < MaxCertificateSize
         if sender_certificate.value.is_some()
-            && sender_certificate.value.as_ref().unwrap().len() >= constants::MAX_CERTIFICATE_LENGTH
+            && sender_certificate.value.as_ref().unwrap().len()
+                >= constants::MAX_CERTIFICATE_LENGTH as usize
         {
             error!("Sender certificate exceeds max certificate size");
             Err(StatusCode::BadDecodingError)
